third_party/WebKit/LayoutTests/fast/dom/Range/compareBoundaryPoints-error.html

<!DOCTYPE html>
<html>
<head>
<title>Document.adoptNode</title>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var r1 = document.createRange();
var r2 = document.createRange();
shouldThrow("r1.compareBoundaryPoints(Range.START_TO_START)", '"TypeError: Failed to execute \'compareBoundaryPoints\' on \'Range\': 2 arguments required, but only 1 present."');
shouldThrow("r1.compareBoundaryPoints(-1, r2)", '"NotSupportedError: Failed to execute \'compareBoundaryPoints\' on \'Range\': The comparison method provided must be one of \'START_TO_START\', \'START_TO_END\', \'END_TO_END\', or \'END_TO_START\'."');
shouldThrow("r1.compareBoundaryPoints(4, r2)", '"NotSupportedError: Failed to execute \'compareBoundaryPoints\' on \'Range\': The comparison method provided must be one of \'START_TO_START\', \'START_TO_END\', \'END_TO_END\', or \'END_TO_START\'."');
</script>
</body>
</html>
